[Cocoa] Web Driver: `RemoteAutomationTarget`s marked for termination and targets that have never been paired may briefly be indistinguishable in application target listing https://bugs.webkit.org/show_bug.cgi?id=240524 <rdar://93430106>
Reviewed by BJ Burg and Devin Rousso.

A `RemoteAutomationTarget` has multiple possible entries it may provide in an application listing to webinspectord,
which are then interpreted by webinspectord to determine the current state of that target. The key pieces for this bug
are that a `WIRConnectionIdentifierKey` which when present is taken to mean that the target is no longer pristine and
`WIRAutomationTargetIsPairedKey` which when true indicates that the target is paired. When a connection identifier is
present and we are no longer paired, the connection is considered terminated.

There exists a race condition where webinspectord preemptively set the targets state to "Terminated" upon sending a
close request in order to make sure that the target is no longer used. Upon receiving the close message for the target
the connection identifier is removed from the map of known connection identifiers before scheduling the actual target to
be torn down asyncronously. In most cases this works fine, but sometimes an application listing is created and sent
between removing the identifier from the map of known connections and actually tearing down the target, and that is
where this issue occurs. We will in that case send a listing for the automation target with no connection identifier
(since we removed it from the map already), which is then interpreted as a listing for a pristine automation target. At
that point, it is possible that automation sessions that are being rapidly created and destroyed will attempt to pair
with the seemingly pristine existing session, even though that session is about to be torn down.

* Source/JavaScriptCore/inspector/remote/RemoteAutomationTarget.h:
* Source/JavaScriptCore/inspector/remote/cocoa/RemoteConnectionToTargetCocoa.mm:
(Inspector::RemoteConnectionToTarget::close):
* Source/JavaScriptCore/inspector/remote/cocoa/RemoteInspectorCocoa.mm:
(Inspector::RemoteInspector::listingForAutomationTarget const):

Canonical link: https://commits.webkit.org/250664@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@294361 268f45cc-cd09-0410-ab3c-d52691b4dbfc