WTF::CrashOnOverflow::crash() with /((a{100000000})*b{2100000000})+/.test(); https://bugs.webkit.org/show_bug.cgi?id=240552
Reviewed by Yusuke Suzuki and Alex Christensen.

Replaced a crashing checked add in the YARR JIT code generator with a reporting add so we can
catch the overflow.  When we do overflow, we terminate the JIT genration with an error.
We'll fallback to the YARR interpreter, which already exits with an error on a similar overflow.

* JSTests/stress/regexp-large-quantifier.js:
(testRegExp):
* Source/JavaScriptCore/yarr/YarrJIT.cpp:
(JSC::Yarr::dumpCompileFailure):
* Source/JavaScriptCore/yarr/YarrJIT.h:

Canonical link: https://commits.webkit.org/250703@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@294411 268f45cc-cd09-0410-ab3c-d52691b4dbfc