[git-webkit] Question 'Are you sure want to setup a fork?' always returns 'Yes'
Reviewed by Jonathan Bedard.

When the response of the user is an empty string, function 'Terminal.choose'
always returns the first option.

The reason for that is a 'startswith' comparison of an option with an empty
string. This comparison is always true, thus the first option is selected
instead of the default option.

Canonical link: https://commits.webkit.org/250790@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@294535 268f45cc-cd09-0410-ab3c-d52691b4dbfc