[Cocoa] fast/text/international/generic-font-family-language-traditional.html is failing due to language minimization https://bugs.webkit.org/show_bug.cgi?id=240751 <rdar://problem/93394618>
Reviewed by Chris Dumez.

We're saying the language list ["en", "zh-tw"] gets minimized to ["en"], which then removes the signal about whether we should render simplified or traditional Chinese.

On one hand, this is kind of good, because the whole point of minimization is to make different browser configurations in the same bucket as identical. On the other hand, it means a user with these language preferences won't get the rendering they expect.

I could change the test to just use the language list ["zh-tw"] and I think it would still test the thing it's trying to test, so that's what this patch does.

* LayoutTests/fast/text/international/generic-font-family-language-traditional.html:
* LayoutTests/platform/ios/TestExpectations:
* LayoutTests/platform/mac/TestExpectations:

Canonical link: https://commits.webkit.org/250878@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@294654 268f45cc-cd09-0410-ab3c-d52691b4dbfc