Add WebKit SPI to specify User Agent details for preconnect https://bugs.webkit.org/show_bug.cgi?id=240818
Reviewed by NOBODY (OOPS!).

Add WebKit SPI to specify application name and whether a desktop user agent should be used in the preconnect.
This is in preparation of fixing the issue where we are always preconnecting with the incorrect user agent on
iPad. This causes the preconnected socket to be unused, and is a page load performance issue.

* Source/WebKit/UIProcess/API/Cocoa/WKWebView.mm:
(-[WKWebView _setUserAgentDetailsForPreconnect:desktopUserAgent:]):
* Source/WebKit/UIProcess/API/Cocoa/WKWebViewPrivate.h:
(-[WKWebView _dataTaskWithRequest:completionHandler:]): Deleted.
* Source/WebKit/UIProcess/Cocoa/WebPageProxyCocoa.mm:
(WebKit::WebPageProxy::setUserAgentDetailsForPreconnect):
* Source/WebKit/UIProcess/WebPageProxy.cpp:
(WebKit::WebPageProxy::preconnectTo):
* Source/WebKit/UIProcess/WebPageProxy.h: