[iPadOS] Align scroll position to origin on resize in Multitasking mode https://bugs.webkit.org/show_bug.cgi?id=240815 rdar://92888214
Reviewed by Devin Rousso.

Make a minor adjustment to avoid re-centering the target unobscured content rect on the hit-tested
node in the center of the viewport, when multitasking mode is enabled; instead, we can simply
preserve the existing unobscured content rect offset.

* Source/WebKit/WebProcess/WebPage/ios/WebPageIOS.mm:

We can also avoid this hit-test when `usesMultitaskingModeViewportBehaviors()` is enabled, since we
only need to anchor to the origin of the current unobscured rect.

(WebKit::WebPage::dynamicViewportSizeUpdate):

Canonical link: https://commits.webkit.org/250894@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@294690 268f45cc-cd09-0410-ab3c-d52691b4dbfc