[iOS] No message displayed when Parental content filter blocked load https://bugs.webkit.org/show_bug.cgi?id=240724
Reviewed by Geoffrey Garen.

The request to unblock a load after it has being blocked by the Parental content filter on iOS needs to be
performed in the UI process, since it may show UI. This path does not set the blocked error description on
the content filter like we do in the other path, which is addressed in this patch. This is a regression
from moving the content filtering to the Network process.

* Source/WebKit/NetworkProcess/NetworkResourceLoader.cpp:
(WebKit::NetworkResourceLoader::contentFilterDidBlock):

Canonical link: https://commits.webkit.org/250931@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@294766 268f45cc-cd09-0410-ab3c-d52691b4dbfc